feat(timestamp): adds timestamp modifier (#75)
* Working through adding a feature to render a timestamp.

* feat: adds support for the timestamp modifier

Adds support for the timestamp modifier which renders an iso8601 timestamp with the log. Also adds
the iso8601 timestamp to the meta data of the log.

* docs: documents the new timestamp modifier

Adds documentation for the new timestamp modifier along with visual examples.

* docs: updated the timestamp documentation in the data section of the guide

* docs: added version information for new timestamp data props

The timestamp data object contains new props of utcTimezoneOffset and iso8601 which were not
available until version >= 1.1. Adding documentation for the version requirements.

* docs: adds version requirements for showTimestamp on data docs

* fix: fixes bug where 0 hour UTC timezone offset should be Z

Co-authored-by: Stacy <ndstc@a483e7d1bd68.ant.amazon.com>