feat(shed): added tools to the shed for easier manual browser debugging
Added a tools namespace to the shed to allow for easier manual browser debugging. This was added
because the shed exists in the global namespace and makes it easy to access when trying to run
commands in the browser console. It mainly just wraps the filter and rendering utility functions
that are exported by the library. It also includes another method that simplifies rendering a
collection of logs from the shed cache."