docs: documented new printed property and the new shed tools (#95)
* feat: logs now always render and printed boolean now returned when terminated

Logs will always render and store their render. To know if a log was printed to the console
terminated logs now expose the "printed" boolean that indicates if it was printed. Log listeners
also now receive the printed boolean. Rather than checking for render equal to null, listeners
should now check the printed boolean.

* feat(shed): added tools to the shed for easier manual browser debugging

Added a tools namespace to the shed to allow for easier manual browser debugging. This was added
because the shed exists in the global namespace and makes it easy to access when trying to run
commands in the browser console. It mainly just wraps the filter and rendering utility functions
that are exported by the library. It also includes another method that simplifies rendering a
collection of logs from the shed cache."

* docs: documented new printed property and the new shed tools

Added documentation for the new printed property on terminated log objects. Also added documentation
for the new $shed.tools.

* docs: added version indicators for the new features to docs

* docs: set the cdn version to always be latest for docs example

Authored-by: Andrew Stacy <ajstacy@gmail.com>