feat(namespace): adds capability to constrain allowed namespaces
Adze factories now can accept a generic type of contraint to restrict the allowed namespaces to a
pre-defined list. This will enable you to control your namespaces centrally and to add a new
namespace you must first add it to the centralized list. This will make understanding what
namespaces are available in your application much clearer.