fix: fixes setup to force configuration. internally uses new getGlobal function
A bug existed where setup configuration would not be set if a log was generated prior to setup being
called. The log would generate a empty configuration and the setup function would return the empty
config rather than setting it. This functionality has been split into two functions to prevent this
behavior.